Move prooftrie to src directory and fix failing import
