+++ packages/trie/src/index.ts
@@ -3,5 +3,6 @@ export * from './db/index.js'
 export * from './node/index.js'
 export * from './proof/index.js'
 export * from './trie.js'
+export * from './prooftrie.js'
 export * from './types.js'
 export * from './util/index.js'
