+++ packages/trie/src/proof/index.ts
@@ -65,4 +65,3 @@ export function verifyTrieRangeProof(
 }
 
 export * from './range.js'
-export * from './proofTrie.js'
